requirements.txt
Appium-Python-Client==2.11.1
pytest==7.4.0
selenium==4.10.0
webdriver-manager==3.8.6
